requirements.txt
OptiCommPy ~= 0.9.0
ipywidgets ~= 8.1.3
numpy ~= 1.24.4
matplotlib ~= 3.9.0
scipy ~= 1.13.1
tqdm ~= 4.66.4
